backend/requirements.txt
# Core FastAPI dependencies
fastapi
uvicorn[standard]
pydantic
python-multipart

# AI and LangChain dependencies
langchain
langchain-community
langchain-google-vertexai
langchain-huggingface
langchain-chroma

# Vector database and embeddings
chromadb
sentence-transformers

# Google Cloud dependencies
google-cloud-aiplatform

# Canvas and LTI dependencies
requests
PyJWT
python-dotenv
sqlalchemy
psycopg2-binary
cryptography

# Utilities
loguru
beautifulsoup4
lxml
cryptography

# Testing
pytest
pytest-asyncio
httpx
requests 

asyncpg